refactor: section table
