--- .eslintignore
@@ -1,31 +1,0 @@
-#####################################################################################
-# Generated
-#####################################################################################
-
-## Packages installed by NPM. https://docs.npmjs.com/about-packages-and-modules
-node_modules/
-
-## Vitest coverage output. https://vitest.dev/guide/coverage.html
-coverage/
-
-## Turbo outputs.
-.turbo/
-
-## Vite build outputs for React.
-build/
-
-## Build outputs from `tsc`, specified by `tsconfig.json`.
-dist/
-
-
-#####################################################################################
-# Development Environment
-#####################################################################################
-
-## Static assets.
-public/
-
-## Private environment variables.
-.env
-.env.*
-!.env.sample
