+++ packages/anteater-api-types/src/index.ts
@@ -2,3 +2,4 @@ export * from './courses';
 export * from './enrollHist';
 export * from './grades';
 export * from './websoc';
+export * from './larc';
